Add the rabbitmq and tomcat scripts

+++ vagrant/Automated_Provisioning/rabbitmq.sh
@@ -2,13 +2,26 @@
 set -euxo pipefail
 
 cd /tmp/
-wget https://packages.erlang-solutions.com/erlang/rpm/centos/7/x86_64/esl-erlang_24.0.2-1~centos~7_amd64.rpm
 
-sudo yum -y install esl-erlang_24.0.2-1~centos~7_amd64.rpm
+## Check if erl(esl-erlang) exists, otherwise download it
+if [ -f "/usr/bin/erl" ]; then
+    echo "Erlang already exists. Skipping download."
+else
+    wget https://packages.erlang-solutions.com/erlang/rpm/centos/7/x86_64/esl-erlang_24.0.2-1~centos~7_amd64.rpm
 
-wget https://github.com/rabbitmq/rabbitmq-server/releases/download/v3.8.19/rabbitmq-server-3.8.19-1.el7.noarch.rpm
+    sudo yum -y install esl-erlang_24.0.2-1~centos~7_amd64.rpm
+fi
 
-sudo yum install -y rabbitmq-server-3.8.19-1.el7.noarch.rpm
+## Install rabbitmq
+if [ -f "/usr/sbin/rabbitmq-server" ]; then
+    echo "Rabbitmq already exists. Skipping download."
+else
+    wget https://github.com/rabbitmq/rabbitmq-server/releases/download/v3.8.19/rabbitmq-server-3.8.19-1.el7.noarch.rpm
+
+    sudo yum install -y rabbitmq-server-3.8.19-1.el7.noarch.rpm
+fi
+
+
 
 ## Install dependencies
 sudo yum install -y socat logrotate
@@ -21,10 +34,19 @@ sudo systemctl enable rabbitmq-server
 echo "[{rabbit, [{loopback_users, []}]}]." > /etc/rabbitmq/rabbitmq.config
 
 ## Creating user with user privileges
-rabbitmqctl add_user test test
-rabbitmqctl set_user_tags test administrator
+user_exists=$(sudo rabbitmqctl list_users | grep "test")
 
-sudo sysetemctl restart rabbitmq-server
+if [[ -z "$user_exists" ]]; then
+    ## User doesn't exist, create it
+    rabbitmqctl add_user test test
+    rabbitmqctl set_user_tags test administrator
+    echo "User test created successfully"
+else
+    ## User already exists, skip creation
+    echo "User test already exists, skipping creation"
+fi
+
+sudo systemctl restart rabbitmq-server
 sudo systemctl status rabbitmq-server
 
 ## Set firewalls for rabbitmq
